Fixed Delete games with looping prompts with no user games

+++ lib/cli.rb
@@ -32,26 +32,38 @@ while option_choice != 5
   elsif option_choice == 1
     add_user_game(current_user)
   elsif option_choice == 2
-    update_user_stats(current_user)
+    if current_user.has_games?
+      update_user_stats(current_user)
+    else
+      no_games
+    end
   elsif option_choice == 3
-    display_option = 7
-    while display_option != 4
-      display_stat_menu
-      display_option = gets.chomp.to_i
-      if display_option == 4
-        puts "Returning to the Menu Main"
-      elsif display_option == 1
-        specifc_game_stats(current_user)
-      elsif display_option == 2
-        average_stats(current_user)
-      elsif display_option == 3
-        lifetime_stats(current_user)
-      else
-        puts "Invalid input"
+    if current_user.has_games?
+      display_option = 7
+      while display_option != 4
+        display_stat_menu
+        display_option = gets.chomp.to_i
+        if display_option == 4
+          puts "Returning to the Menu Main"
+        elsif display_option == 1
+          specifc_game_stats(current_user)
+        elsif display_option == 2
+          average_stats(current_user)
+        elsif display_option == 3
+          lifetime_stats(current_user)
+        else
+          puts "Invalid input"
+        end
       end
+    else
+      no_games
     end
   elsif option_choice == 4
-    delete_game(current_user)
+    if current_user.has_games?
+      delete_game(current_user)
+    else
+      no_games
+    end
   else
     puts "Invalid input"
   end
